refactor: extract path helpers & error handling to lib/sp/helpers.ts (Phase 3)
